+++ src/components/EmptyResultsWithIcon/index.tsx
@@ -1,0 +1,29 @@
+/*
+ * Copyright (c) 2021 AccelByte Inc. All Rights Reserved.
+ * This is licensed software from AccelByte Inc, for limitations
+ * and restrictions contact your company contract manager.
+ */
+
+import * as React from "react";
+import "./index.scss";
+import { default as classNames } from "classnames";
+
+export interface EmptyResultsProps {
+  imgSrc: string;
+  title: string;
+  description?: string | React.ReactNode;
+  children?: React.ReactChild;
+  dataQa?: string | null;
+  isTaller?: boolean;
+}
+
+export const EmptyResultsWithIcon = ({ imgSrc, title, description, children, isTaller, dataQa }: EmptyResultsProps) => {
+  return (
+    <div className={classNames("emptyResultsWithIcon", { taller: isTaller })} data-qa-id={dataQa}>
+      <img src={imgSrc} />
+      <div className="emptyResultsWithIconTitle mt-2">{title}</div>
+      {description && <div className="emptyResultsWithIconDescription">{description}</div>}
+      {children && <div>{children}</div>}
+    </div>
+  );
+};