+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [0.1.34]
 ### Added
 - Publish the components storybook to github page
+- Add more stories file to the components that don't have it yet
 
 ## [0.1.33]
 ### Added